Use the Shazam v2/detect endpoint
We don't need any of the new features which seem to be based around making a smarter attempt at multiple attempts of identifying a song.

Due to the nature of the recorder, there wouldn't be a good way to integrate that into the recorder.

But, the v1 endpoint is deprecated so may as well switch.